+++ tests/pools/worker-node.test.mjs
@@ -58,6 +58,7 @@ describe('Worker node test suite', () => {
         executing: 0,
         queued: 0,
         maxQueued: 0,
+        sequentiallyStolen: 0,
         stolen: 0,
         failed: 0
       },
@@ -84,7 +85,6 @@ describe('Worker node test suite', () => {
       threadWorkerNode.tasksQueue.size
     )
     expect(threadWorkerNode.onBackPressureStarted).toBe(false)
-    expect(threadWorkerNode.onEmptyQueueCount).toBe(0)
     expect(threadWorkerNode.taskFunctionsUsage).toBeInstanceOf(Map)
 
     expect(clusterWorkerNode).toBeInstanceOf(WorkerNode)
@@ -101,6 +101,7 @@ describe('Worker node test suite', () => {
         executing: 0,
         queued: 0,
         maxQueued: 0,
+        sequentiallyStolen: 0,
         stolen: 0,
         failed: 0
       },
@@ -127,7 +128,6 @@ describe('Worker node test suite', () => {
       clusterWorkerNode.tasksQueue.size
     )
     expect(clusterWorkerNode.onBackPressureStarted).toBe(false)
-    expect(clusterWorkerNode.onEmptyQueueCount).toBe(0)
     expect(clusterWorkerNode.taskFunctionsUsage).toBeInstanceOf(Map)
   })
 
@@ -155,6 +155,7 @@ describe('Worker node test suite', () => {
         executed: 0,
         executing: 0,
         queued: 0,
+        sequentiallyStolen: 0,
         stolen: 0,
         failed: 0
       },
@@ -178,6 +179,7 @@ describe('Worker node test suite', () => {
         executed: 0,
         executing: 0,
         queued: 0,
+        sequentiallyStolen: 0,
         stolen: 0,
         failed: 0
       },
@@ -201,6 +203,7 @@ describe('Worker node test suite', () => {
         executed: 0,
         executing: 0,
         queued: 0,
+        sequentiallyStolen: 0,
         stolen: 0,
         failed: 0
       },